Logfile works, dryrun doesn't

+++ get_v0.py
@@ -7,7 +7,7 @@ import numpy as np
 import random
 from itertools import combinations, cycle, product
 from collections import defaultdict, deque
-from scaffold import Scaffold, Longreads, LongReadSVG, revcomp
+from scaffold import Longreads, LongReadSVG, revcomp
 from statistics import mean
 
 parser = ArgumentParser()
@@ -16,10 +16,15 @@ parser.add_argument("sequencefile", help="Input File in FASTQ format")
 parser.add_argument("contigstringfile", help="Input File containing subsequent longread ids")
 parser.add_argument("contigfile", help="Contig File in FASTA format")
 parser.add_argument("--blacklistfile", help="File containing long read ids where certain contig mappings should be ignored.")
+parser.add_argument("--dryrun", default=False, action="store_true", help="Do not output sequence to file, but only print the length of the resulting sequence to stdout.")
+parser.add_argument("--logfile", help="Logging File")
 parser.add_argument("linename", help="Name of cell line")
 parser.add_argument("outfile", help="Sequence Output File")
 args = parser.parse_args()
 
+seq1="CCCTTTTTGGGAAA"
+print(seq1)
+print(revcomp(seq1))
 contigs = {}
 for read in SeqIO.parse(args.contigfile, "fasta"):
     contigs[read.id] = str(read.seq)
@@ -34,9 +39,13 @@ if args.blacklistfile:
             else:
                 blacklist[sline[0]].append(sline[1])
 
-lreadseqs = {}
-for read in SeqIO.parse(args.sequencefile, "fastq"):
-    lreadseqs[read.id] = str(read.seq)
+if not args.dryrun:
+    lreadseqs = {}
+    for read in SeqIO.parse(args.sequencefile, "fastq"):
+        lreadseqs[read.id] = str(read.seq)
+
+if args.logfile:
+    loghandle = open(args.logfile,"w+")
 
 print("Nr. of Contigs: " + str(len(contigs)))
 
@@ -44,8 +53,12 @@ scafs = Longreads(args.inputfiles, blacklist, args.linename)
 #scafs.filter_small_contigs(300)
 #scafs.filter_reverse_small_contigs(600)
 scafs.filter_low_quality_contigs(0.76)
-scafs.turn_longreads_around()
+print(scafs.lreads["88e9f524-98be-4630-98bd-06054fd07355"]["maps"][10])
+scafs.turn_longreads_around(logging=loghandle)
+print(scafs.lreads["88e9f524-98be-4630-98bd-06054fd07355"]["maps"][10])
+#print(scafs.lreads["88e9f524-98be-4630-98bd-06054fd07355"])
 scafs.sort_by_starts()
+scafs.filter_contigs_by_coverage(0.5,ignore_ends=True, verbose = False)
 
 print("Nr. of reads: " + str(len(scafs.lreads)))
 
@@ -61,32 +74,39 @@ print("Nr. of reads: " + str(len(scafs.lreads)))
 #        print("Could not find first contig")
 #        sys.exit()
             
+seqlength = 0
 
 
 items = []
 with open(args.contigstringfile) as f:
     for line in f:
+        if line.startswith("#"):
+            continue
         if len(line.split()) == 2:
+        
             lrid, contig = line.split()
             if lrid not in scafs.lreads:
                 print("Error while reading " + args.contigstringfile)
                 print("Did not load " + str(lrid) + " from the input files")
                 sys.exit()
-            if lrid not in lreadseqs:
-                print("Error while reading " + args.contigstringfile)
-                print("Did not load sequence for " + str(lrid))
-                sys.exit()
-            ctgn = contig + args.linename
+            if not args.dryrun:
+                if lrid not in lreadseqs:
+                    print("Error while reading " + args.contigstringfile)
+                    print("Did not load sequence for " + str(lrid))
+                    sys.exit()
+            #ctgn = contig + args.linename
+            ctgn = contig
             if ctgn not in contigs:
                 print("Error while reading " + args.contigstringfile)
                 print("Contig " + str(contig) + " not found.")
                 sys.exit()
             items.append([lrid, ctgn])
         elif len(line.split()) == 1:
-            ctgn = line.rstrip() + args.linename
+            #ctgn = line.rstrip() + args.linename
+            ctgn = line.rstrip() 
             if ctgn not in contigs:
                 print("Error while reading " + args.contigstringfile)
-                print("Contig " + str(contig) + " not found.")
+                print("Contig " + str(ctgn) + " not found.")
                 sys.exit()
             items.append([ctgn])
 
@@ -97,6 +117,8 @@ with open(args.contigstringfile) as f:
 
 
 for item in items:
+    if args.dryrun:
+        break
     if len(item) == 2:
         if item[0] not in lreadseqs:
             print(str(item[0]) +" not found in sequence file. Aborting !")
@@ -108,9 +130,16 @@ distances[("476APD","1405APD")] = -9
 distances[("2316APD","1498APD")] = -12
 distances[("504APD","49APD")] = -447
 distances[("849APD","1471APD")] = -11
-firstcontig = "2345APD"
+distances[("1192DBB","462DBB")] = -9712
+#firstcontig = "2345APD"
+#firstcontig = "901DBB"
+#firstcontig = "1003SSTO"
+firstcontig = items[0][0]
+print("First contig: " + firstcontig)
+items = items[1:]
 
 out_sequence = contigs[firstcontig]
+overall_length = len(out_sequence)
 last_ctgn = firstcontig
 
 # check that important contigs weren't filtered out
@@ -124,28 +153,30 @@ for item1, item2 in zip(items[:-1], items[1:]):
             print("Problem found.")
             print(ctgn + " not in " + lrid2)
             sys.exit()
+if args.dryrun:
+    print("Getting length of sequence ...")
+else:
+    print("Getting sequence ...")
 
-print("Getting sequence ...")
+longN = "N"*2000000
 
 for item in items:
     #print("length of sequence: " + str(len(out_sequence)))
     #print(item)
     if len(item) == 2:
-        if last_ctgn == "1471APD":
-            print(item)
         first_ctgn = last_ctgn
         lrid, last_ctgn = item
-        if last_ctgn == "1391APD":
-            print(item)
-        if "reverse" in scafs.lreads[lrid]:
-            #print(lrid + " is reverse complimentary")
-            lr_seq = revcomp(lreadseqs[lrid])
+        if args.dryrun:
+            lrseq = longN
         else:
-            lr_seq = lreadseqs[lrid]
+            if "reverse" in scafs.lreads[lrid]:
+                #print(lrid + " is reverse complimentary")
+                lr_seq = revcomp(lreadseqs[lrid])
+            else:
+                lr_seq = lreadseqs[lrid]
+        
         status = 0
         for ctg in scafs.lreads[lrid]["maps"]: # they should be ordered
-            if last_ctgn == "1391APD":
-                print(ctg)
             #print("ctgn: " + ctg["name"])
             if ctg["strand"] == 1: # no reverse contigs allowed
                 continue
@@ -157,16 +188,15 @@ for item in items:
                 else:
                     continue
             elif status == 1:
-                tocut = len(contigs[last_used_ctg["name"]]) - (last_used_ctg["ecc"]+1) # ecc and scc are zero based
+                #tocut = len(contigs[last_used_ctg["name"]]) - (last_used_ctg["ecc"]+1) # ecc and scc are zero based
                 #print("to cut: " +str(tocut) + " " + last_used_ctg["name"])
-                if tocut > 0:
-                    out_sequence = out_sequence[:-tocut]
+                #if tocut > 0:
+                #    out_sequence = out_sequence[:-tocut]
                 if last_used_ctg["ecr"] > ctg["scr"]:
                     out_sequence = out_sequence[:ctg["scr"]-last_used_ctg["ecr"]]
-                    out_sequence += contigs[ctg["name"]][ctg["scc"]:]
                 else:
                     out_sequence += lr_seq[last_used_ctg["ecr"]+1:ctg["scr"]]
-                    out_sequence += contigs[ctg["name"]][ctg["scc"]:]
+                out_sequence += contigs[ctg["name"]][ctg["scc"]:ctg["ecc"]]
                 if ctg["name"] == last_ctgn:
                     break
                 last_used_ctg = ctg
@@ -182,12 +212,14 @@ for item in items:
                 out_sequence = out_sequence[:distances[(last_ctgn, item[0])]]
                 out_sequence += contigs[item[0]]
                 last_ctgn = item[0]
-                
-            
-with open(args.outfile, "w") as out:
-    out.write(">MHC_APD\n")
-    out.write(out_sequence)
 
-        
-    
+if args.logfile:
+    loghandle.close()
+                
+if args.dryrun:
+    print("Lenght of the assembly: " + str(len(out_sequence)))
+else:
+    with open(args.outfile, "w") as out:
+        out.write(">MHC_APD\n")
+        out.write(out_sequence)
